Move param "editMode" to store.

+++ lyrebird_bugit/attachment.py
@@ -41,6 +41,8 @@ def remove_attach():
 
 def rename(attachment_item, new_file_name):
     raw_file_path = Path(attachment_item.get('path'))
+    if not raw_file_path.exists():
+        raise FileNotFoundError
     new_file_path = raw_file_path.parent.joinpath(new_file_name)
     os.rename(raw_file_path, new_file_path)
     attachment_item['path'] = str(new_file_path)
